apps/api/pyproject.toml
[project]
name = "z-image-api"
version = "0.1.0"
description = "Z-Image backend API (FastAPI + Celery)"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "celery>=5.4.0",
    "redis>=5.0.0",
    "psycopg[binary]>=3.2.0",
    "python-dotenv>=1.0.0",
    "boto3>=1.35.0",
    "modelscope>=1.32.0",
    "huggingface-hub>=1.1.5",
]
